requirements.txt
# Core Python dependencies for DeepSea-AI Stage 1 (Preprocessing & Vectorization)
# Pinned minimal versions; adjust as needed
pandas>=2.2.0
numpy>=1.26.0
biopython>=1.83
scikit-learn>=1.4.0
pyyaml>=6.0.1
click>=8.1.7
loguru>=0.7.2
rich>=13.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.33.0
plotly>=5.20.0
# For running R from Python
rpy2>=3.5.11
# Optional for progress bars
tqdm>=4.66.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-dotenv>=1.0.1
pydantic>=2.6.0
httpx>=0.27.0
orjson>=3.9.15
uvloop; platform_system != 'Windows'
pytest>=8.1.0
fastapi-cli>=0.0.4
python-multipart>=0.0.9
